+++ layer-admin/src/main/java/org/layer/space/controller/dto/AdminSpaceCountGetResponse.java
@@ -1,4 +1,4 @@
-package org.layer.domain.admin.controller.dto;
+package org.layer.space.controller.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
