+++ layer-external/src/main/java/org/layer/oauth/service/OAuthService.java
@@ -4,5 +4,4 @@ import org.layer.oauth.dto.service.MemberInfoServiceResponse;
 
 public interface OAuthService {
     MemberInfoServiceResponse getMemberInfo(final String accessToken);
-
 }
